+++ src/OFS/interfaces.py
@@ -597,8 +597,6 @@ class ICopyContainer(Interface):
         argument.
         """
 
-    manage_renameForm = Attribute("""Rename management view""")
-
     def manage_renameObjects(ids=[], new_ids=[], REQUEST=None):
         """Rename several sub-objects"""