+++ CHANGES.rst
@@ -32,7 +32,7 @@ Restructuring
 
 - Remove property management ZMI screens.
 
-- Remove ZMI copy/cut/paste and re-ordering features.
+- Remove ZMI copy/cut/paste/rename and re-ordering features.
 
 - Drop `OFS.History` functionality.